// moonbowcl/src/app/model/dto/criterio-ricerca-istanze.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */


export class CriterioRicercaIstanze {
  cfDichiarante: string;
  nomeDichiarante: string;
  cognomeDichiarante: string;
  stato: number;
  importanza: number;
  codiceIstanza: string;
  idModulo: number;
  titoloModulo: string;
  dataDa: Date;
  dataA: Date;
  inviateDataDa: Date;
  inviateDataA: Date;
  sort: string;
  nomeElementoIstanza: string;
  valoreElementoIstanza: string;
  protocollo: string;
  idMultiEnte: number;
  filtroEpay: string;
}